fix bug in cross validation routines using sklearn cross validated linear models.
Previous use was wrong model.score was used but this has nothing to do with cross validation error. Instead now extract cv score from model attributes. Unfortunately omp does not store the needed information even though it is computed. For now identified a patch that can be applied by the user. Also turned of lars as it does not exit when regularization parameter alpha strats to increase. This exit condition is used by sklearn for lasso but nor lars for some reason