Fix failing building of cython modules.
The code stopped compiling overnight and I could not find the cause of the change. The pull request of the last commit to devel passed all tests then when I pulled this devel and even old branches the cython modules no longer built. This occured on several machines. It looks like the line was ctypedef np.int_t int_t was causing the issue. I just changed all instances of int_t to int64_t